Let a series volume fall between two numbers
Some series number their in-between books 5.6, and a few open with a
0.5 before the first volume. The column was a smallint, so Postgres
rounded such a volume to the nearest whole one without a word; it is
numeric(5, 2) now, and two decimals is also where the form rounds.

The field keeps its own text while it is being typed. Parsing every
keystroke straight into a number would have swallowed the separator:
after "5," the value reads back as 5, the field rewrites itself, and
the comma can never be typed at all. It becomes a number on save.
Comma and point are both accepted, the keyboard is asked for a decimal
one, and the detail page says "Band 5,6" the German way round.

Zero still means no volume, while 0.5 passes - that is precisely how
the prequel to the first book gets numbered.

+++ supabase/schema.sql
@@ -9,7 +9,7 @@ create table public.books (
   subtitle text,
   authors text[] not null default '{}',
   series text,
-  series_volume smallint,
+  series_volume numeric(5, 2),
   isbn text,
   published_year smallint,
   page_count integer check (page_count > 0),
